Major Calibration Update
Added calibration routine tweaks that allow for the final C4 note to also be calibrated by the end user. In normal playing mode, the octaves are limited to C0 through C3 with the extra button (High C) playing C4 which was originally hard coded. This version adds calibration for the high C4, by allowing a user to use the octave up button to go up one additional octave in calibration mode only, making the C4 available for calibration on the lower C key. This is necessary because the High C button in calibration mode is used as a "shift key" to confirm editing the values of the key selected.