Experimental Branch Version V1.01-2C
This version replaces the 5V trigger with a 5V Gate. 
It should retain all the calibration functions (although testing beyond jumping in and out of calibration mode is yet to be done) 

Using the gate adds a level of complexity to the project in the form of additional code to keep the gate from closing on the release of one button when two buttons are accidentally pressed together, debouncing, and additional functions surrounding the high c key to make sure the gate is steady on that note, and not looping in a fashion that causes an envelope to retrigger.   

The code has also been tweaked to remove blockers that could cause the gate to wait for the slew to complete before the gate can be triggered. 

It remains to be seen if this functionality is critical to the success of the module, and if the hardware limitations of the platform can manage a smooth user experience. If I can get it working well enough, I will pull it into the main branch. If not, I may explain the tradeoffs and let the end users pick their firmware.