Update SlightofHand_v1_01EXPERIMENTAL2H.ino
Many changes. Working as expected now with one bug that I am working on. When the portamento is on the slew works normally on all keys EXCEPT the High C key. When you play a slew note from any normal key, up to the High C, that note will "stick" and continue to play until another note is pressed, or that High C note is hit again. I believe this is due to the slew code implementation is in a blocking fashion. I'm working on it, but it's getting closer.