Stability Update - Experimental
This includes the rewrite to include a 5V gate instead of a trigger. Debouncing logic has been imported, and gate function is now running stable. Slew bug where slewing up to the high c would cause it to stick has been fixed. This may be ready to be pulled into the Main branch shortly.